feat: 도메인 엔티티 설계
* 대표 감정
Diary(N) - (1)Emoji
* 오늘의 감정
Diary(N) - (M)Emoji
=> Diary(1) - (N)Today_Emoji(M) - Emoji(1)

//TODO : Account - Diary 관계 매핑

issue #5
