+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,6 @@
+#Wed Mar 10 00:40:21 KST 2021
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.8.2-all.zip
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.8.2-bin.zip
-zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
