+++ src/main/java/dev/solar/daymoji/domain/diary/Diary.java
@@ -1,26 +1,45 @@
-package dev.solar.daymoji.model.diary;
+package dev.solar.daymoji.domain.diary;
 
 import lombok.*;
 
+import javax.persistence.*;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
 @Getter
-@EqualsAndHashCode(of = "id")
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
+@Entity
 public class Diary {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String title;
+
     private String contents;
+
+    @ManyToOne
+    @JoinColumn(name = "EMOJI_ID")
     private Emoji representativeEmoji;
+
+    @OneToMany(mappedBy = "diary")
+    private List<TodayEmoji> todayEmojis = new ArrayList<>();
+
     private double latitude;
+
     private double longitude;
+
     private String nameOfLocation;
+
     private boolean isOpened;
-    private LocalDateTime createdAt;
-    private LocalDateTime modifiedAt;
+
+    private LocalDateTime createdDate;
+
+    private LocalDateTime modifiedDate;
 
     public void modifyTitle(String title) {
         this.title = title;
