+++ src/main/java/dev/solar/daymoji/domain/diary/EmojiRepository.java
@@ -1,0 +1,6 @@
+package dev.solar.daymoji.domain.diary;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EmojiRepository extends JpaRepository<Emoji, Long> {
+}
