+++ src/main/java/dev/solar/daymoji/domain/diary/TodayEmojiRepository.java
@@ -1,0 +1,6 @@
+package dev.solar.daymoji.domain.diary;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TodayEmojiRepository extends JpaRepository<TodayEmoji, Long> {
+}
