+++ src/main/resources/application.properties
@@ -8,6 +8,12 @@ spring.datasource.url=ENC(iwVSR6WZ4ZbtmVtR8xZ/I83BLebg/SR8tSjtNknyvoA5GGtAkU+2O8
 spring.datasource.username=ENC(mOdSnamebcjq/X7leytVSw==)
 spring.datasource.password=ENC(l9rg74YEDjX/yrAHDS0aWg==)
 
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+spring.jpa.properties.hibernate.format_sql=true
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+
 #logging
 logging.level.dev.solar.daymoji=DEBUG
 logging.level.sql=DEBUG
